engine: Adding information about aggregate link status to Bond
Bonds in mode 4 (link aggregation) have an aggregate assigned
when working properly. This mac is reported by vdsm. This was
added by patch:
https://gerrit.ovirt.org/#/c/53100/14
This patch adds the aggregate mac as property of Bond, and saves
it to DB so it can be later used in the UI to alert the user
about a possible error condition.

Change-Id: I96e9d6f0dfeeaa98becb0768cff44d30f038154b
Bug-Url: https://bugzilla.redhat.com/1281666
Signed-off-by: mirecki <mmirecki@redhat.com>
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>