core: prevent host in maintenance while vm runs on it in db
Before switching a host to maintenance, we set its status to prepare-to-
maintenance and migrate the VMs that run on it. Only when no VMs run on
the host, we switch it to maintenance.

The problem is that we count on the number of VMs that are reported by
the host. It is racy because we might get this report from VDSM before
the Down event arrived or after the Down event has been ignored.

This patch refine the check that is done before switching a host to
maintenance to check that no VM is running on the host in the database
as well. We check the database only if VDSM reports that no VM runs on
it so it should be fine performance wise.

Change-Id: Ia953d71bb0b036ed0e1e65527c650e955a031fa1
Bug-Url: https://bugzilla.redhat.com/1325938
Signed-off-by: Arik Hadas <ahadas@redhat.com>