restapi: Fix system permissions server
In version 3 of the engine the system permissions resource was
implemented using an empty interface as a placeholder for the required
JAX-RS annotations. But in version 4 of the engine the classes that
implement the V3 compatibility layer don't use inheritance, and as a
result this resource is actually empty. This patch fixes that, removing
the empty class and replacing it with the generic assigned permissions
implementation.

Change-Id: I2f8d5ac0e0107538e4aae917b1f71fade0f670f4
Bug-Url: https://bugzilla.redhat.com/1334096
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>