core: commands in EXECUTION_FAILED status might have running childs
When the command status is EXECUTION_FAILED the command might
still have running child commands, there command shouldn't be
considered as ended until all of its child commands were ended.

Currently not all callbacks has the appropiate handling to that
status and its sometimes assumed that a child command on that status
was ended with failure.

This patch adds pollOnExecutionFailed() to indicate whether doPolling()
should be executed for the command if its on EXECUTION_FAILED status
untill all the callbacks will inheirt from the same base callback.

Change-Id: I713b4d179316e056466d402fe3201fad5666b849
Signed-off-by: Liron Aravot <laravot@redhat.com>