restapi: Don't require admin/filter to access root service
Currently access to the root service requires admin privileges or the
use of the "Filter" header. If a non privileged users tries to use this
service without the "Filter" header an error is returned because some of
the queries that it uses fail. To avoid that this patch changes the
implementation of this service so that it will explicitly check the
result of the queries. If a query fails the corresponding data won't be
added to the returned result, but the overall request will succeed.

Change-Id: I6b8582238dede59724277795c32640ee1bef808f
Bug-Url: https://bugzilla.redhat.com/1323631
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>