engine: disable IPv6 out-of-sync reporting
Currently VDSM reports "dhcpv6=true" when an IPv6 address was assigned
dynamically. Similar logic worked for IPv4 where the only method to get
the address dynamically is DHCP indeed. However, IPv6 address could be
assigned dynamically by the stateless autoconfiguration protocol too,
which makes the described logic not reliable.
The described VDSM bug causes the engine marking a network as
out-of-sync when the "autoconf" boot protocol was requested -
VDSM applies the requested correctly, but GetCaps reports that
erroneously.

Reporting out-of-sync IPv6 configuration is disabled temporary.
It's planned to be re-enabled after v4.0-beta is released.

Change-Id: I5d3e8de70e24f21ede03a875da4de2526145d033
Signed-off-by: Yevgeny Zaspitsky <yzaspits@redhat.com>