Add the filter into the "Host Device" detail information.
Add the field useFilterList. If useFilterList is true, the filter
will work. Use the BLACK list to filter the Host Device information,
say, these won't display on the frontend. The BLACK list consist of
regex expression. The white is a ensurance, if the BLACK list filter
the items which you don't want.Use the WHITE list ensure these display.
The BLACK list consist of regex expression, too.

Use the W representes the match WHITE list with items collection,
the B representes the match BLACK list with items collection, the result
is (W||(!B)).

Signed-off-by: GoLine <sline798@gmail.com>