restapi: Manually add RSDL links impossible to generate
Some of the links that should be part of the RSDL are currently
impossible to generate automatically. The reason is that some of the
JAX-RS interfaces are generic, and it is impossible for the RSDL
generator to automatically infer what methods are available in the
implementations of those interfaces. In the master branch this has been
solved as part of the introduction of the new model and metamodel, but
doing the same in the 3.6 branch is too complicated, and risky. To solve
the issue in a less risky way this patch adds a new "rsdl_links.xml"
file that contains manually generated links that will be added to the
RSDL during the build process.

Change-Id: I57a0ba2bbdf9ed4d72bb1c1f15b8af274799f092
Bug-Url: https://bugzilla.redhat.com/1331068
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>