restapi: Fix system errata server
In version 3 of the engine the system errata resource was implemented
using an empty interface as a placeholder for the required JAX-RS
annotations. But in version 4 of the engine the classes that implement
the V3 compatibility layer don't use inheritance, and as a result this
resource is actually empty. This patch fixes that, removing the empty
class and replacing it with the generic errata implementation.

Change-Id: I1ef667d5e9a0e698a7f4bafb7a7843e61ced019d
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>