core: dedicated query for monitored network interfaces
Previously, we used GetVmNetworkInterfaceViewByVmId that uses the
costly vm_interface_view in order to query the network interfaces
whose statistics are being monitored.

This patches adds a new procedure that uses a much lighter view
than vm_interface_view that will be used by the monitoring instead.

There is still a lot of place for improvement but this change
seems to be the simpler and one that would probably be needed
along with further improvements anyway as it reduces the time to
fetch the data by more than four times.

Change-Id: I475ffe644c7636e1e3a5ddea474d08f4b6ee80bd
Signed-off-by: Arik Hadas <ahadas@redhat.com>