core: fix NPE on import template
A regression that was caused by recent refactoring lead to a NPE
in the validation of ImportVmTemplateCommand. The problem was
that the ImportValidator was instanciated using its empty ctor,
thus the ID of the storage-pool was null when reaching the
validation of mac-addresses.

This patch fixes this by passing the parameters of the
ImportVmTemplateCommand that contain the ID of the storage-pool
to the ImportValidator.

Change-Id: I133ceca9cb78a3c7bb67a967d9baba53065e7ea4
Bug-Url: https://bugzilla.redhat.com/1336405
Signed-off-by: Arik Hadas <ahadas@redhat.com>