core: Change singletonList to new Array for StorageCommandBase.
Each command inherited from StorageDomainCommandBase,
which does not implement the procedure getPermissionCheckSubjects,
will risk an exception of UnsupportedOperationException
once a new permission will be added to it.

An example of a failure can be found when using enabling
quota in a Data Center.

Hence, changing permissionCheckSubjects singleton collection
to use a new array list instead.

Change-Id: Ida67a4409509652d5eeb1a953c3537051e4e61bb
Signed-off-by: Maor Lipchuk <maorlipchuk@gmail.com>