core: Don't send existingConnAction in VM ticket
We currently send existingConnAction parameter in each VM ticket, which
prevents the user from overriding this parameter (e.g. in Vdsm hooks).
This is not what we want.  Additionally, sending the parameter is
unnecessary, so we should stop sending it.  If we don't send the
parameter, libvirt/QEMU should behave in the same way as with
"disconnect" value, while we don't prevent the user from altering the
value in Vdsm hooks.

An even better solution would be to allow using different values for
existingConnAction ("disconnect" and "keep") in Engine and send them to
Vdsm in the creation path.  However, this is a bit complicated on the
Vdsm side as using "keep" requires setting qemu-cmdline options in
domain XML, which is not what we want to do now.  So we just implement
this workaround and let the user play with multiple SPICE screens using
means such as Vdsm hooks.

Change-Id: I800d31bc5a28b7cb896cbb11ea0441e07cfbf26e
Bug-Url: https://bugzilla.redhat.com/1060573
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>