userportal, webadmin: move constant and message text to properties files
With GWT i18n, there is the capability to define the default text of a
constant or message as an annotation on the corresponding i18n interface
method definition.  In order to better integrate with the translation
tool we use (Zanata), all default English text needs to be contained in
property files.

A source analysis and transform tool was run to:
  1. extract the default/source English text from annotations
  2. put the text into a new properties file
  3. update existing localized versions of the properties file
     to remove any key not defined in the source properties file
     (this removed inherited duplicate keys)

When a new key is added to a Constants or Messages interface in future,
the default text will need to be defined in a properties file instead
of a source code annotation.

Change-Id: I32cb3ad52beb84dbf3523c8290aede4fc1c5d385
Bug-Url: https://bugzilla.redhat.com/1224423
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>