core: SuperUser can not use virtio serial console
The problem was that the query returning the list of VMs where the user has the
CONNECT_TO_SERIAL_CONSOLE used the GetAllFromVmsForUserAndActionGroup stored
procedure.

This stored procedure was based on the user_vm_permissions_view which did not
contain the admin roles.

Fixed by creating a new view vm_permissions_view which contains both the admin
and user roles.

Change-Id: Ia5f07ab7fd89714b4ab3709639c6e15dfdba1d8e
Bug-Url: https://bugzilla.redhat.com/1320343
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>