core: Map PM iLO3 and iLO4 to their native agents
IPMI is currently implicitly used when working with iLO3/4 fence
devices.
This was done from historical reasons because fence-agents package did
not had specific agents for iLO3/4
But that was changed and now we can use iLO3/4 native drivers which are
recommended.
Also, lanplus=1 is set by default for the native drivers and there is no
need to set it implicitly anymore.

Patch was tested with iLO4 agent and works as expected.
I will try to find a way to validate it on iLO3 as well and also on PPC
machines.

Change-Id: I6f5b5799d5f394f194a26edd6ab54172ef4c2205
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1126753
Signed-off-by: emesika <emesika@redhat.com>