core: RunVm - do not pass compensation context
When running stateless vm, the RunVm command first creates a snapshot for the vm.
When running the create snapshot command, the command context is passed to the child
command - which means that the command compensation context is passed as well.
The compensation in use is NoOp compensation context it causes the
executed child command to use that compensation context as well, on case
of failure the db changes won't be reverted which will leave us with
inconsistent db state.

Change-Id: I93a6035b3a0691a0b1640444e49ef92b128d1122
Bug-Url: https://bugzilla.redhat.com/1334105
Signed-off-by: Liron Aravot <laravot@redhat.com>