webadmin: v2v-source vms list is cleaned in failure
While opening the Webadmin Import VM dialog and importing from
VMware source or any other external source (v2v), the "virtual
machines on source" list was not cleaned in case of entering
invalid parameters.

After this fix: if you enter valid parameters and click the
"load" button -> the "virtual machines on source" list is displayed.
If you then enter incorrect parameters and click
the "load" button again (either field validation failure or
connection failure), the "virtual machines on
source" list is now cleaned while an error is displayed.

The bug was opened for VMware but this fix is done for
all external sources.

Change-Id: I85d5e415a34ad4a92c42bd6763ebadb7698f3372
Bug-Url: https://bugzilla.redhat.com/1298293
Signed-off-by: Sharon Gratch <sgratch@redhat.com>