core: Avoid an infinite loop when untar an entry
Avoid going into infinite loop in case there is an error during an entry
extraction from a tar archive.
For example if the tar contains a corrupted or simply an empty file,
this will trigger the infinite loop. The fix will just skip to the next
entry and will add empty files into the archive (we reflect the tar
entries and not ignoring them)

Change-Id: I64a8d019cb5c13df19ee21eecb2ab81b42a9b3cc
Bug-Url: https://bugzilla.redhat.com/1167262
Signed-off-by: Roy Golan <rgolan@redhat.com>