tools: Add AlertOnNumberOfLVs to engine-config
Add the ability to control this config value via engine-config.sh, as
different storage arrays in different users' environments may be less or
more susceptible to the slowness associated with a large number of LVs.

This patch does not introduce a limitation on the maximal value for this
config value (setting it to Integer.MAX_INT) as to not introduce any
behavior changes in upgraded systems where a user may have set it to
some arbitrary large value manually and would not be able to use the
tool to achieve the same behavior.
It does, however, impose a minimum value of 6. This is the number of
special LVs created when a block storage domain is created, and setting
the config value to anything lower than that would be pointless.

Change-Id: I302d99dc09ba8fce81e8b6ff6bd040bfc14fd8d0
Bug-Url: https://bugzilla.redhat.com/1338545
Signed-off-by: Allon Mureinik <amureini@redhat.com>