webadmin: Add label comboboxes to VM & host dialogs
This patch adds read-only comboboxes to the general tab of the
add/edit dialog boxes for VMs and hosts. The boxes are
populated with all available affinity labels. For existing VMs
and hosts, labels that have been set through the REST API are
shown as selected.

This is phase 1 of this feature. Phase 2 will introduce the
ability to change selected labels via the UI. Until that time,
labels can be modified via the REST API.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1338799
Change-Id: Iad36512b423a20b3ad7ad9e01ff43882555fcdb6
Signed-off-by: Phillip Bailey <phbailey@redhat.com>