engine: Add/UpdateVmInterfaceCommand fix compensation
'VmBase.ssoMethod' field caused the bug.
The enum 'SsoMethod' overrides 'toString()' method.
Since 'JsonObjectDeserializerFactory' uses
'EnumResolver.constructUnsafeUsingToString' to create a map of enum keys
to enum values, the key of the map is the value of the 'toString()' -> the
lower case name of the enum.
On the other hand, the serializer uses the 'name()' method as the key of
the serialized enum. Since in 'SsoMethod' the 'name()' and the
'toString()' are not the same, the deserialier cannot find the serialized
enum.

Added '@JsonCreator' and '@JsonValue' to make sure the key in both
serialization and deserialization will be the same.

Change-Id: Iad3a12d1542e16ca7d8aa6c8a2c142bb2e09e9fd
Bug-Url: https://bugzilla.redhat.com/1338601
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>