core: ChildCommandsCallbackBase - setSucceeded(false)
When the command entity is persisted, its return value is persisted as well.
If the command execute() succeeded, getReturnValue().getSucceeded() will return
true.

The command object is built from the persisted command entity, which means
that the return value will be initiated to the persisted return value, when calling
endAction() having the return value returning true value when checking
getSucceeded() might lead to wrong results.

This patch is simple and sets the succeeded flag to false to avoid that issue.

Change-Id: I39fd64001b0ef5354853aa3e6433d478f8dbb74a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1339536
Signed-off-by: Liron Aravot <laravot@redhat.com>