engine: Adjusting the value of binding:host_id during openstack port allocation
The value of binding:host_id must be the same value as was set during the automated
neutron agent during host installation, or just the host name if the agent was
installed manually. If the neutron agent was installed automatically during host
deployment, the vds static data will contain the provider id. Based on this the
value the host_id is determined.

Change-Id: I6f0fd4c2398112c189b853c1294bf535b82dca9b
Signed-off-by: mirecki <mmirecki@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1318543