core: Make the download size of HE config disk configurable
The size of the config disk to download is configurable and can be
changed depending underlying storage infra. Since the config tar doesn't
occupy the 'real' disk size that is reported we can safely download the
exact section that holds the tar image.

TODO the size of the tar should be parsed from the description, much
like it is used today for OVF disks. See
https://bugzilla.redhat.com/show_bug.cgi?id=1336655

Change-Id: I7541c8372b044370346e20127ebca434a9245781
Bug-Url: https://bugzilla.redhat.com/1167262
Signed-off-by: Roy Golan <rgolan@redhat.com>