restapi: Add V3 IO exception mapper
Currently version 3 of the API doesn't have a specific IO exception
mapper, as exceptions are handled by version 4, and then tranlsated to
the version 3 format. But some exceptions happen before version 4 is
called, for example the exceptions that happen during validation of
integer values. In order to handle this kind of exceptions like we used
to do in the past, we need a version 3 specific mapper. This patch adds
it.

Change-Id: I2622a162bfd92bc29bb35271e918c9cc7f3c8b6f
Bug-Url: https://bugzilla.redhat.com/1339907
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 9c23fd98783402d559ab156810ee5f445c9aa48c)