restapi: Adding 'type' to openstacknetworkprovider v3 in  adapter
An openstack network provider can not be created using rest api v3.
The problem is the type property which has been added.
When the v3 provider is translated to v4, this property must be
added. The property must always be set to NEUTRON, as this is the
only provider type supported before v4.

Change-Id: I63c35a3f9823c199fc00b08a0d17d0a05aa56d33
Bug-Url: https://bugzilla.redhat.com/1340862
Signed-off-by: mirecki <mmirecki@redhat.com>