restapi: Add support for V3 SD delete with body
Version 3 of the API supported the use of a body in the operation that
deletes a storage domain. This has been removed from version 4 of the
API, as bodies are in general forbidden for delete operations. But we
still need to support it for version 3 of the API, as that is how the
"format", "destroy" and "host" parameters are handled. This patch
modifies the backwards compatibility layer so that it accepts a body in
this case, and transforms the parameters into query parameters.

Change-Id: I37746b316f284854aab19674c0f9ab4d74cfbf9f
Bug-Url: https://bugzilla.redhat.com/1340164
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>