webadmin: v2v -moving vms between lists got stuck
when choosing vms to import in the import VMs dialog, then clicking
"Next" to next import dialog for choosing cluster and then clicking
"Back" to display the first import dialog again:
in this senario when moving vms from importedVmModels list to
externalVmModels and vice versa, at some point clicking the arrow
button does nothing (arrow button is clicked and button callback is
called but VM is not moved).

This problem occurred because when clicking "back" button and
displaying 1st dialog, new UI objects are created for the 1st dialog
(SplitTable,ImportVmsPopupView...), but the old ones for this 1st dialog
are still alive and both old and new UI objects share the same model.

This causes the old UI button callbacks to disrupt the data for the new
UI objects. in some point when clicking on
a vm to select it with checkbox, the "old" button callback calulated it
as 2 vms checked although only one vm is actually checked.
Trying to move this VM by clicking the arrow causes failure and nothing
is moved.

The fix for this bug includes reseting both importedVmModels and
externalVmModels lists of the model (copying the items list only)
before creating the new UI instances for the "Back" operation.

Change-Id: Ib34d108a2959a50df0e39edd21360d2e5feb4f54
Bug-Url: https://bugzilla.redhat.com/1338740
Signed-off-by: Sharon Gratch <sgratch@redhat.com>