core: fix monitoring of guest agent nics
We used to to compare the hash that is being calculated by the engine (of
network interfaces reported by the guest agent) with the hash in the
dynamic VM data reported by VDSM. The problem is that VDSM doesn't report
this hash, therefore it is always zero and therefore we update the dynamic
data of a VM that is installed with guest agent on each monitoring cycle.

This patch fixes this by changing the mentioned check so the computed
hash will be compared to the one that was calculated previously.

Change-Id: I38533ac778ffef528340d33cfbb64589d5d49fe8
Bug-Url: https://bugzilla.redhat.com/1366786
Signed-off-by: Arik Hadas <ahadas@redhat.com>