webadmin: Populate external network id during subnet create
AddNetworkOnProviderCommand requires external network id
to create subnets on an external provider. Only the network
id, but not the external id was returned after the network
was created.
The patch changes the return value after the network was
created from just the network id to network, so that all
required parameters can be retrieved.

Change-Id: I5941abd9148f7d6dce45c7056ef7224e7ec51904
Bug-Url: https://bugzilla.redhat.com/1216081
Signed-off-by: mirecki <mmirecki@redhat.com>