core: fix inconsistency with empty cd-rom value
By default, the CD-ROM attached to the VM is NULL.
VDSM also reports NULL value in case no CD-ROM is attached.

Recently RunVmCommand was changed to persist the CD-ROM
that is attached to a VM as well, however, it is then set
to an empty string if no CD-ROM is attached. Therefore, the
monitoring gets empty string from the DB and NULL value
from VDSM and decides to update VM dynamic data because of
this change. But later on, the NULL value from VDSM is
ignored and thus, this flow will happen on every monitoring
cycle, leading to many redundant updates of VM dynamic daa.

Change-Id: I5e2afff36f3ee02b57a43760ab88f0a9c3176f37
Bug-Url: https://bugzilla.redhat.com/1340722
Signed-off-by: Arik Hadas <ahadas@redhat.com>