core: move guest memory related fields to statistics
The information related to the memory within the guest used to be
stored in VmDynamic. This is wrong. First, conceptually it is more
'statistics' then 'runtime' info. Second, it is updated frequenty
and therefore leads to many updates of VM dynamic data by the VMs
monitoring.

This patch moves these fields to VmStatistics instead.

Change-Id: If9944576a3283d77ad6c46c14f366efdf536b816
Bug-Url: https://bugzilla.redhat.com/1340722
Signed-off-by: Arik Hadas <ahadas@redhat.com>