core: Use the hosted storage domain id instead of name when importing it
When importing hosted storage domain the system relied on a hardcoded
name: "hosted_storage".
Instead of using that constant to find and import the domain the code is
now using the domain id that is reported by vdsm in the vm data.
The system verifies that a hosted engine VM is running on a storage
domain to determine if that domain is a hosted engine domain.

Change-Id: Iad7c5f4e5557e4c861e5a2317c79bc3fae874e70
Bug-Url: https://bugzilla.redhat.com/1301105
Signed-off-by: Jenny Tokar <jtokar@redhat.com>
(cherry picked from commit 31d8036)