core: cleanup in vm fetchers
VmsStatisticsFetcher only needs to override the poll request from VDSM
(GetAllVmStats instead of List) - it doesn't need to override the long
VmsListFetcher#fetch for that.

In this patch only the relevant part from 'fetch' method that needs to
be overridden is extracted to separate method. This change reduce code
duplication.

In addition, add SupressWarning on intentional unchecked cast and
rename a variable for better readability.

Change-Id: I3eeb9ebaa04e5c52de8cc6b03162054239d5ddc6
Signed-off-by: Arik Hadas <ahadas@redhat.com>