Adding dao tests to std ci
Basically this patch is all about running dao tests
inside a mock. For this, a postgresql DB should be
installed and run inside the mock.

Added postgres packages to .packages files

Created a new script, dao-tests.sh. This script is
being called only if files that are relevant to DB
had been changed. The script will pick a port
randomly and will update pg_hba.conf file to allow
DB access.

All the rest is the same as how the tests were
running when not in std ci.

Change-Id: If01639ebd0e346e0bd724a3b9ec68b44e7e9e765
Signed-off-by: Gil Shinar <gshinar@redhat.com>