webadmin: support sort by client and server side in one table
Before when we have table we could sort only server side or client side,
depends on flag we set. There could be situations where we need to have
support for both. An example is 'Active User Sessions' tab, where we show
columns which are in database and also in memory.

This patch add support, so we can have some columns sorted server side,
and some by client side.

Change-Id: Ic4c1dba19cedf015c5cad42a4f6872a79669d944
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1340451
Signed-off-by: Ondra Machacek <omachace@redhat.com>