core: Remove old shared status calculation
Commit 904ab61 introduced an improved method of calculating a storage
domain's shared status, which is indeed currently in use in the
dwh_storage_domain_history_view.

That patch, however, left the old mechanism in place alongside the new
mechanism. This had no functional effect, as it's a left join on a
column that's never queried, but it does cause a redundant performance
penalty.

This patch fixes the problem by removing this redundant join.

Related-To: https://bugzilla.redhat.com/1306325
Change-Id: I835cf5ccbedba2c54c1b216ef6bb267a40f89a75
Signed-off-by: Allon Mureinik <amureini@redhat.com>