core: Fix creation of diskless snapshot without memory
When creating a snapshot without memory the snapshot is created
initially in OK status while relying on the fact that no async
operations will run and the command won't reach the endAction()
phase.

When using the ConcurrentChildCommandsExecutionCallback, the
command will always reach the endAction() method.
This patch removes duplicated code between the execute() and the
endAction() while adapting the endAction() to handle a snapshots
which contains only configuration.

Change-Id: Ica11d72fa366a65eda975770ad96d7ec1feb8701
Bug-Url: https://bugzilla.redhat.com/1342110
Signed-off-by: Liron Aravot <laravot@redhat.com>