core: Correct flow of RunVmCommand.endAction()
Recent change in RunVmCommand.getCallback() forces RunVmCommand to
actively use the callback. This requires changes in other places to
correctly execute RunVmCommand.endAction() in every situation.

The flow selected by RunVmCommand.getFlow() method is now cached in
RunVmParams. The cache in RunVmCommand itself is useless, because the
command is created from scratch when CoCo calls its callback, and
getFlow() cannot determine the flow correctly in this situation.

AttachUserToVmFromPoolAndRunCommand is modified to correctly use the
callback - without it endAction() is never called. RunVmCommand is
started as COMMAND_MANAGED and performs by itself all actions needed to
start a VM as stateless or stateful.
AttachUserToVmFromPoolAndRunCommand.endAction() is executed after that
and checks the final result of RunVmCommand. If it failed, user's
ownership of the VM is reverted.

Change-Id: I099d5a8b7ee3aef6b8abb124de39767de6a99c20
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1342795
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>