core: Monitor action executions with hystrix
Wrap all action executions into synchronous hystrix commands.
This allows easy live monitoring of all actions.

The hystrix event stream servlet is listening on

  http://<host>/ovirt-engine/services/hystrix.stream

Hystrix monitoring is disabled by default. It first needs to be enabled with

   engine-setup -s HystrixMonitoringEnabled=true

and a restart of the engine is required.

Afterwards it can be controlled with JMX MXBean methods:

  HystrixCommandController#monitor*('true')

Also more fine grained control on which commands should be monitored is
possible in the JMX bean.

To get a nice monitoring dashboard use hystrix-dashboard and point it to
the hystrix stream url. The url is protected with basic auth, like the
rest api. Hystrix dashboard requires the basic auth parameters to be
provided in encoded form. If your credentials are admin@internal:engine
you have to provide 'Basic YWRtaW5AaW50ZXJuYWw6ZW5naW5l'

Benefits are:

 * Easier verification while implementing features, because failing
   commands are immediately visible on the dashboard.
 * Easier verification for QA because of the same reason.
 * Immediate insights in what is going on inside the application for our
   solution architects and support when customers are facing problems.

Bug-Url: https://bugzilla.redhat.com/1347628
Change-Id: Ie2d70707b6f5d72fdcf8457ea74fad30c75d7170
Signed-off-by: Roman Mohr <rmohr@redhat.com>