engine: prevent notification flood for number of LVs
When creating more than <AlertOnNumberOfLVs> LVs on a storage domain
(300 by default), we log that the number of LVs on the storage domain
exceeded the threshold.

Since one can register to get this notification by mail, we should
prevent a mail flood.
This patch restricts this event to be logged once in 3 hours for each
storage domain.

Change-Id: I4df27e5b7211cb00c49221e99d4cdeab82f21bd9
Bug-Url: https://bugzilla.redhat.com/1275182
Signed-off-by: Idan Shaby <ishaby@redhat.com>