core: fix detection of unmanaged vms
When detecting an unmanaged VM running on the host, we add the dynamic
data that is reported by VDSM to the database. This data does not contain
the host that the VM is running on, so we end up having a running VM
without the host that it runs on. This situation must be prevented.

Change-Id: I9f488df83c9335922b0b422e85da4bc93b8a380b
Bug-Url: https://bugzilla.redhat.com/1342818
Signed-off-by: Arik Hadas <ahadas@redhat.com>