engine: Can't assign ip for a network attached to the same nic as mgmt
The validation that caused this wrong blocking is
'networkIpAddressWasSameAsHostnameAndChanged()'.
Comparing the code of the validation in the old setup network command
to the new one, shows that besides the bug in the title, the old logic
was changed by mistake.
After rethinking about this validation we came to conclusion that the
old and the new validations block cases that may be permitted by the vdsm.
So we decided, for now, to remove the engine level validation. In case of
a real problem, the vdsm will perform a rollback.

Change-Id: Ia965a98c0a466d77238f45254f95e008145befcc
Bug-Url: https://bugzilla.redhat.com/1340702
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>