webadmin: display invalid partner mac alert also when partner mac is 00:00:00:00:00:00
The invalid partner mac alert on mode 4 bonds (exclamation
mark next to bond name) must be displayed not only when the
mac is empty, but also when the mac is 00:00:00:00:00:00.

The partner mac is empty or 00:00:00:00:00:00 when there is
an error condition in the bond (mode=4). This is why we
must guard against this specific mac value.

Change-Id: I13bc65d4097edb7b84e1c77a1f539d2f24800019
Bug-Url: https://bugzilla.redhat.com/1281666
Signed-off-by: mirecki <mmirecki@redhat.com>