core: db deadlock by disk image dynamic updates
VmsMonitoring updates the disk image dynamic table with
the dynamic image data retrieved from the vm.
Different engine flows can be executed when the vm is up, when
the different flows end its attempted to complete the image data
to the engine.

VmsMonitoring performs batch update on sorted collection, while currently
the updates done by the different flows that inheirt from BaseImagesCommand
aren't sorted - as both updates may be executed in transaction we might
reach a deadlock.

This patch solves the deadlock by performing the image dynamic update out of
transaction when its performed from the code in BaseImagesCommand which should solve
the issue. This updates just sets the size retrieved from the storage, so we shouldn't
care if its commited even of the rest of the changes are rolled back.

Change-Id: Ie1645bd1f6bc991f9ec8501ad8aa8dd7d8987ee5
Bug-Url: https://bugzilla.redhat.com/1342317
Signed-off-by: Liron Aravot <laravot@redhat.com>