core: HE - uploadImageToStream task is never cleared
When downloading the data from the HE configuration disk the created task
isn't polled which means it'll remain forever.
When attempting to stop the spm its first checked whether there are pending
tasks. as the task will remain forever it won't be possible to stop the spm.

This patch changes the code to use the bll command instead of the vds command,
which will take care of polling and clearing the task.

Change-Id: I978be73b8df88a59bd30a0f114285d5999e353f7
Bug-Url: https://bugzilla.redhat.com/1343618
Signed-off-by: Liron Aravot <laravot@redhat.com>