backend: generate default template's name for repo-image
When importing an image from glance as a template, if the template's
name is not specified then we generate a default one in
ImportRepoImageCopyTaskHandler::createTemplate.

The problem with this is that ImportRepoImageCommand doesn't know about
this auto generated template's name (it's null in the command's
parameters), and then the template is logged as a part of the event as
"<UNKNOWN>", although it's created with its default name.

This patch generates that default template's name in the command instead
of the handler.

Change-Id: I7386be3d2c1254dc3df0dfab9c6c340b41a00aed
Bug-Url: https://bugzilla.redhat.com/1324780
Signed-off-by: Idan Shaby <ishaby@redhat.com>