restapi: don't ignore new template's name on import image from Glance
When importing an image as a template from Glance, if the URL that is
used is of the form
"/api/openstackimageproviders/<provider_id>/images/<image_id>/import"
and the template's name is specified, it is ignored and not passed to
ImportRepoImageCommand.

This patch passes this parameter if it is specified.

Change-Id: I60f44c0d519e7deea29095543b1a06ac0d6a1ef4
Bug-Url: https://bugzilla.redhat.com/1344048
Signed-off-by: Idan Shaby <ishaby@redhat.com>