core: Auto select correct graphics device on Template creation
The treatment performed in AddVmCommand initilization was missing from
AddVmTemplateCommand. Moved the method to VmHandler and called both from
AddVmCommand and AddVmTemplateCommand.

Also when importing to ppc cluster another bug manifested itself
with auto selection of soundcard. Since soundcard is not supported on ppc
the null default value passed in parameters was not overwritten and
caused NPE in later command execution. Now we set the soundcard
explicitly to false if it remains null after the treatment.

Change-Id: I66a4b8175f13511df2ed95cff5893522a18a9e4b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1341639
Signed-off-by: Martin Betak <mbetak@redhat.com>