core: DiskImageDynamicDao - deadlock caused by wrong sorting
updateAllDiskImageDynamicWithDiskIdByVmId is supposed to sort
the DiskImageDynamicDao objects its updating in order to avoid
deadlocks. Currently the sorting is done on the first object
of the pair which is the vm id instead of the the image group id.

This patch changes the sorting to be according the image group id as
it was supposed to be.

Change-Id: I4194c93b0e919543b5a5c1b3621dab0d4604c983
Bug-Url: https://bugzilla.redhat.com/1342317
Signed-off-by: Liron Aravot <laravot@redhat.com>