restapi: Adapt "Link" header for V3
Version 4 of the API doesn't return the "capabilities" link because it
doesn't exist. Version 3 of the API adds it explicitly, but only to the
response body, it doesn't add it to the "Link" header. This patch
modifies the V3 compatibility layer so that the "Link" header generated
by version 4 of the API is removed and replaced by a value calculated
from the request body, this way it will the body and the header will
always be in sync.

Change-Id: Ieb2ad00a84e671629d8303898c7b08cd300cded2
Bug-Url: https://bugzilla.redhat.com/1342226
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit bef31bbe3c628e059e09d63d04540ae7deafeef7)