packaging: Don't explicitly disable usage of JBoss modules indexes
JBoss modules has the capability to read and write module indexes stored
in external files. This used to be enabled by default till version 1.2.0
of JBoss modules. When it was enabled by default it caused potential
problems for the engine:

* The engine could try to write indexes in directories where it doesn't
  have permission to write.

* The engine could try to use indexes that had been written for older
  versions of the application server.

To avoid these two problems the engine was modified to explicitly
disable writing the modules, and to create a shallow copy of the modules
directories without the existing indexes.

But since version 1.2.0 of JBoss modules both reading and writing these
external index files is disabled by default. The version of the
application server that we use uses version 1.5.0 of JBoss modules, so
we no longer need any of these two workarounds. This patch removes them
completely.

Change-Id: I1aa9d9091e634be9d7154120d95bd6ae7a4142c1
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>